+++ packages/tests/server/___testHelpers.ts
@@ -1,6 +1,7 @@
 import type { IncomingMessage } from 'http';
 import http from 'http';
 import type { AddressInfo, Socket } from 'net';
+import { waitError } from '@trpc/server/__tests__/waitError';
 import type { TRPCWebSocketClient, WebSocketClientOptions } from '@trpc/client';
 import {
   createTRPCClient,
@@ -171,47 +172,6 @@ export function routerToServerAndClientNew<TRouter extends AnyRouter>(
 
 export async function waitMs(ms: number) {
   await new Promise<void>((resolve) => setTimeout(resolve, ms));
-}
-
-type Constructor<T extends object = object> = new (...args: any[]) => T;
-
-export async function waitError<TError extends Error = Error>(
-  /**
-   * Function callback or promise that you expect will throw
-   */
-  fnOrPromise: Promise<unknown> | (() => unknown),
-  /**
-   * Force error constructor to be of specific type
-   * @default Error
-   **/
-  errorConstructor?: Constructor<TError>,
-): Promise<TError> {
-  let res;
-  try {
-    if (typeof fnOrPromise === 'function') {
-      res = await fnOrPromise();
-    } else {
-      res = await fnOrPromise;
-    }
-  } catch (cause) {
-    // needs to be instanceof Error or DOMException
-    if (
-      cause instanceof Error === false &&
-      cause instanceof DOMException === false
-    ) {
-      throw new Error(
-        'Expected function to throw an error, but it threw something else',
-      );
-    }
-    if (errorConstructor) {
-      expect((cause as Error).name).toBe(errorConstructor.name);
-    }
-    return cause as TError;
-  }
-
-  // eslint-disable-next-line no-console
-  console.warn('Expected function to throw, but it did not. Result:', res);
-  throw new Error('Function did not throw');
 }
 
 export async function waitTRPCClientError<TRoot extends InferrableClientTypes>(
