+++ packages/tests/server/getRawInput.test.ts
@@ -1,4 +1,4 @@
-import { waitError } from './___testHelpers';
+import { waitError } from '@trpc/server/__tests__/waitError';
 import type { TRPCError } from '@trpc/server';
 import { initTRPC } from '@trpc/server';
